refactor: move Device classes into Toy namespace
